feat(boss/v2): Expand list of known apps
Use the JSON list of apps documented on the archival-tools repo as
source for these apps. The title IDs were obtained by querying
tasksheets of these apps, but not all of them were available so the
title ID is set to 0 in those cases.

The title name and region were populated though various public sources,
but they aren't perfect so some names and/or regions are left unknown.